DE4490 Make PayPal Express Checkout Function
PayPal express don't work when the backend setting
for the PayPal express payment method
`Transfer Cart Line Items` is turn off. The set
Express checkout payload is configured to validate the
following fields `LineItemsTotal`, `ShippingTotal`,
and `TaxTotal`. However, these fields are not
populated with data from the quote when the
`Transfer Cart Line Items` setting is turn off.
Also, the `<LineItems>` node is optional in the
set Express Checkout payload, it is still being
sent when the `Transfer Cart Line Items` setting
is turn off causing no `Order Total` to be displayed
on the PayPal page.