Static Payload Config Map
Use a locator object for looking up data on how to construct payloads.
Use the locator in the PayloadFactory instead of directly depending upon
the configuration in PayloadConfigMap.php. Load the configuration in
PayloadConfigMap into one type of locator, using a static property to
prevent the mapping from being loaded into memory more than once. This
should dramatically reduce memory usage of the SDK in cases where
multiple instances of the PayloadFactory are used - e.g. any payloads
with subpayloads.